Fix SpecialKey linking issue for :dig visibility
When using diacritics the colorscheme makes it hard to read the digraph
symbols this is due to the `hi! link SpecialKey Whitespace' setting that
is implemented when not in nvim. I don't know the reason for this being
in so I added a configuration item to tweak this behavior. It is fully
backwards compatible.

Closes: #114

Signed-off-by: Wesley Schwengle <wesleys@opperschaap.net>